pyproject.toml
[tool.poetry]
name = "dataspace_backend"
version = "0.1.0"
description = "Proof of Business backend"
authors = ["George J Padayatti <george.padayatti@igrant.io>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "3.0.7"
django-auditlog = "1.0.0"
djangorestframework = "3.13.1"
djangorestframework-simplejwt = "4.3.0"
django-cors-headers = "3.7.0"
psycopg2-binary = "2.8.6"
PyJWT = "1.7.1"
gunicorn = "^20.1.0"
django-filter = "21.1"
django-rest-auth = "0.9.5"
django-constance = "2.9.1"
django-picklefield = "3.0"
django-jsonfield = "^1.4.1"

[tool.poetry.dev-dependencies]
ruff = "^0.11.0"

[tool.poetry.group.dev.dependencies]
mypy = {version = ">=1.0.0", python = ">=3.9"}
django-stubs = {extras = ["compatible-mypy"], version = "^5.2.9", python = ">=3.10"}
djangorestframework-stubs = {extras = ["compatible-mypy"], version = "^3.16.7", python = ">=3.10"}
types-requests = {version = "^2.32.4.20260107", python = ">=3.9"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["config", "onboard", "organisation", "service", "data_disclosure_agreement", "data_disclosure_agreement_record", "connection", "webhook", "software_statement", "notification", "b2b_connection", "oAuth2Clients", "discovery", "authorization", "customadminsite", "dataspace_backend"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.10"
strict = true
exclude = "migrations/"

# Ignore missing stubs for third-party Django packages
[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = [
    "django.*",
    "rest_framework.*",
    "django_filters.*",
    "auditlog.*",
    "rest_framework_simplejwt.*",
    "constance.*",
    "picklefield.*",
    "jsonfield.*",
    "rest_auth.*",
    "corsheaders.*",
]
ignore_missing_imports = true
